Modified branch to follow the architecture, keep some consistency
